Move to desktop directory remote Browserstack desktop platforms
