+++ lib/browsenator/remote/browserstack/desktop/screen_size.rb
