+++ spec/browsenator/remote/browserstack/desktop/chrome_spec.rb
@@ -1,11 +1,11 @@
-describe Browsenator::Remote::Browserstack::Chrome do
+describe Browsenator::Remote::Browserstack::Desktop::Chrome do
   describe '#open' do
     after(:each) do
       @browser.quit
     end
 
     it 'starts Chrome version 66 in a Mac when no version is specified' do
-      @browser = Browsenator::Remote::Browserstack::Chrome.new(project: 'Test').open
+      @browser = described_class.new(project: 'Test').open
       browser_type = @browser.driver.capabilities.browser_name
       browser_version = @browser.driver.capabilities.version
       operating_system = @browser.driver.capabilities.platform
@@ -18,8 +18,8 @@ describe Browsenator::Remote::Browserstack::Chrome do
     end
 
     it 'starts Chrome version 65 in Mac when that version is specified' do
-      @browser = Browsenator::Remote::Browserstack::Chrome.new(project: 'Test', browser_version: '65.0',
-                                                               screen_width: 1280, screen_height: 960).open
+      @browser = described_class.new(project: 'Test', browser_version: '65.0',
+                                     screen_width: 1280, screen_height: 960).open
 
       browser_type = @browser.driver.capabilities.browser_name
       browser_version = @browser.driver.capabilities.version
