+++ spec/browsenator/remote/browserstack/desktop/edge_spec.rb
@@ -1,11 +1,11 @@
-describe Browsenator::Remote::Browserstack::Edge do
+describe Browsenator::Remote::Browserstack::Desktop::Edge do
   describe '#open' do
     after(:each) do
       @browser.quit
     end
 
     it 'starts Edge' do
-      @browser = Browsenator::Remote::Browserstack::Edge.new(project: 'Test').open
+      @browser = described_class.new(project: 'Test').open
       browser_type = @browser.driver.capabilities.browser_name
 
       expect(@browser).to be_a(Watir::Browser)
