+++ spec/browsenator/remote/browserstack/desktop/safari_spec.rb
@@ -1,11 +1,11 @@
-describe Browsenator::Remote::Browserstack::Safari do
+describe Browsenator::Remote::Browserstack::Desktop::Safari do
   describe '#open' do
     after(:each) do
       @browser.quit
     end
 
     it 'starts Safari version 11.1 in a Mac when no version is specified' do
-      @browser = Browsenator::Remote::Browserstack::Safari.new(project: 'Test').open
+      @browser = described_class.new(project: 'Test').open
       browser_type = @browser.driver.capabilities.browser_name
       browser_version = @browser.driver.capabilities.version
       operating_system = @browser.driver.capabilities.platform
@@ -18,7 +18,7 @@ describe Browsenator::Remote::Browserstack::Safari do
     end
 
     it 'starts Safari version 10.1 in Mac when that version is specified' do
-      @browser = Browsenator::Remote::Browserstack::Safari.new(project: 'Test', browser_version: '10.1').open
+      @browser = described_class.new(project: 'Test', browser_version: '10.1').open
       browser_type = @browser.driver.capabilities.browser_name
       browser_version = @browser.driver.capabilities.version
       operating_system = @browser.driver.capabilities.platform
